Prevent circular inclusion
Check "include" directives for circular dependencies and raise
a warning (similar to the one rised when file-insertion is disabled).

To prevent endless loops, we
rise a warning when a nested inclusion with the same "path",
"first-line", "end-line", "start-after", and "end-before"
options is detected.
In order to achieve this, the current source's parents are logged.

Subversion-Revision: r8562