HTML5: add self-link after section headings (feature request #28).
The "section_self_link" setting triggers this behaviour. As it is
currently only implemented for HTML5, the setting is defined for this
writer only.

The self-link "invisible" by default. The provided stylesheet
"responsive.css" gives an example how it can be made visible when
hovering over the section heading.

Subversion-Revision: r8788