Move ``quicktest`` to ``tools/dev``
The quicktest tool has not been substantively modified since 2006.
It may still be useful for generating test-cases, so move it to the
developer tools section and do not advertise it as a front-end tool.
Update documentation where relevant to remove mention of quicktest
as a front-end tool, and remove obsolete "EasyDialogs" code path.

Subversion-Revision: r9201