Config setting "root-prefix": configurable root directory for included files.
Cf. [feature-requests:#91]

The "root_prefix" setting is also applied when a writer converts an image
URI to a local filesystem path.
(The image file may be red by the writer to get the image size (HTML
writers) or to embed it in the output (HTML5 and ODT writers).)

Subversion-Revision: r9492