LaTeX to MathML: Fix "mathvariant" handling for numbers.
Conversion of \mathtt{0.12} did not use monospaced digits.

Do a per-character lookup/replacement with
"alphanumical mathematical chars" for the "text" of `<mn>`.

While identifiers are usually put in separate `<mi>` elements
and more than one letter in a `<mi>` indicates a function name
that should not change fonts,
numbers in `<mn>` consisting of several digits should be
changed according to the given "mathvariant".

Subversion-Revision: r9534