Make effect of centre-aligning figures visible in functional HTML text.
Horizontal figure alignment is irrelevant if the figure is full-width.
The figure content is left-aligned with the default styles.

Change the functional test input by specifying 40% figure width to the
example of a centre-aligned figure.

Note: figure alignment is ignored with LaTeX output, figures are
centre-aligned by default (depends on LaTeX packages/documentclass).

Subversion-Revision: r9643